Exception and validation

+++ src/main/java/com/client/crud/services/ClientService.java
@@ -3,7 +3,11 @@ package com.client.crud.services;
 import com.client.crud.dto.ClientDTO;
 import com.client.crud.entities.Client;
 import com.client.crud.repositories.ClientRepository;
+import com.client.crud.services.exceptions.DatabaseException;
+import com.client.crud.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -19,8 +23,9 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById (Long id){
-        Optional<Client> client = repository.findById(id);
-        ClientDTO dto = new ClientDTO(client.get());
+        Optional<Client> result = repository.findById(id);
+        Client client = result.orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
+        ClientDTO dto = new ClientDTO(client);
         return dto;
     }
 
@@ -40,15 +45,26 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto){
-        Client entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
     @Transactional
     public void deleteById(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
